#277 - extracting inner class HackernewsItemFactory
